KNOX-1842 - Upgrade httpclient to 4.5.10
HttpClient 4.5.7 broke url normalization. Knox
didn't have any tests for this case and so we
had to revert after the fact. HttpClient 4.5.8
fixed a lot of the url normalization and some
libraries decided to turn url normalization off.

This commit does the following:
* Adds a test for %2F - KNOX-1005
    * This test passes under HttpClient 4.5.6 and 4.5.8+
    * It breaks as expected under HttpClient 4.5.7
* Adds an explicit config enabling url normalization
    * Ensures that we are in control of url normalization
    * Adds a test for this configuration as well
* Test with both HttpClient normalization enabled and disabled
    * `rest-assured` doesn't expose `RequestConfig` to disable
url normalization
    * Shows how to use HttpClient in `GatewayBasicFuncTest`

All the url safe characters like %2F are fixed by HTTPCLIENT-1968.

The case of `/abc///def` is normalized to `/abc/def` the same
way that Knox does internally with `getPathInfo` and Java URI.

Signed-off-by: Kevin Risden <krisden@apache.org>