KNOX-2354 An HBASEJARS service which can proxy HBase jars hosted by the HBASEUI
In HBase, we can host a "mini" maven repo in hbase itself which a maven
application can point to.

We want this to be in its own service because it has different semantics
than the normal hbaseUI service has (e.g. one would be hooked up to SSO
login, but SSO won't work against that), and Maven cannot function with
the query-string based backend-routing that we do to make the HBaseUI
look like a normal website (instead of a collection of independent
webservers).