KNOX-2376 Ensure all HBASEJARS IN rules are for /hbase/jars and not /hbase/maven
Looks like I was getting turned around the first time, confusing what is
an internal path (HBase UI) and what is an external path (Knox gateway path).
This fixes the service such that we're always coming in to Knox at /hbase/jars
and always going "out" (to hbase ui) as /static/maven.

Further, because we expect this to be deployed as an API (rather than a
UI), we can do away with the rules to render the quirky jetty-dir.css.