KNOX-2839 - Identity assertion provider handles Hadoop ProxyUser auth using the 'doAs' query parameter
With this change, the previously added ProxyUser configuration for KnoxToken impersonation is removed becuase the same can be achieved with the appropriate identity-assertion configuration.

Another very important change is that Knox's HadoopAuth filter no longer sets the 'doAs' user as primary principal. Instead, it's passed down as an ImpersonatedPrincipal and the original user (real user) remains the PrimaryPrincipal in the underlying security context (Subject).